chore: 테스트 픽스처를 베이스에 추가할 수 있게 하기 위해서 기본 페키지를 추가.
