+++ src/main/java/com/giwankim/core/web/filter/CharacterEncodingFilter.java
@@ -1,0 +1,32 @@
+package com.giwankim.core.web.filter;
+
+import java.io.IOException;
+import javax.servlet.Filter;
+import javax.servlet.FilterChain;
+import javax.servlet.FilterConfig;
+import javax.servlet.ServletException;
+import javax.servlet.ServletRequest;
+import javax.servlet.ServletResponse;
+import javax.servlet.annotation.WebFilter;
+
+@WebFilter(value = "/*")
+public class CharacterEncodingFilter implements Filter {
+
+  private static final String DEFAULT_ENCODING = "UTF-8";
+
+  @Override
+  public void doFilter(ServletRequest request, ServletResponse response, FilterChain filterChain)
+      throws IOException, ServletException {
+    request.setCharacterEncoding(DEFAULT_ENCODING);
+    response.setCharacterEncoding(DEFAULT_ENCODING);
+    filterChain.doFilter(request, response);
+  }
+
+  @Override
+  public void init(FilterConfig filterConfig) throws ServletException {
+  }
+
+  @Override
+  public void destroy() {
+  }
+}
