+++ src/main/java/com/giwankim/next/controller/UserNotFoundException.java
@@ -1,0 +1,19 @@
+package com.giwankim.next.controller;
+
+public class UserNotFoundException extends RuntimeException {
+
+  public UserNotFoundException() {
+  }
+
+  public UserNotFoundException(String message) {
+    super(message);
+  }
+
+  public UserNotFoundException(String message, Throwable cause) {
+    super(message, cause);
+  }
+
+  public UserNotFoundException(Throwable cause) {
+    super(cause);
+  }
+}
