--- src/main/java/core/mvc/Controller.java
@@ -1,10 +1,0 @@
-package core.mvc;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-public interface Controller {
-  String execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException;
-}
