--- src/main/java/next/controller/HomeController.java
@@ -1,18 +1,0 @@
-package next.controller;
-
-import core.db.Database;
-import core.mvc.Controller;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-public class HomeController implements Controller {
-
-  @Override
-  public String execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-    request.setAttribute("users", Database.findAll());
-    return "/index.jsp";
-  }
-}
