--- src/main/java/next/controller/LogoutController.java
@@ -1,20 +1,0 @@
-package next.controller;
-
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-import java.io.IOException;
-
-@WebServlet(value = "/user/logout")
-public class LogoutController extends HttpServlet {
-
-  @Override
-  protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-    HttpSession session = req.getSession();
-    session.removeAttribute(UserSessionUtils.SESSION_USER_KEY);
-    resp.sendRedirect("/");
-  }
-}
