--- src/main/java/next/support/context/ContextLoaderListener.java
@@ -1,31 +1,0 @@
-package next.support.context;
-
-import core.jdbc.ConnectionManager;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.jdbc.datasource.init.DatabasePopulatorUtils;
-import org.springframework.jdbc.datasource.init.ResourceDatabasePopulator;
-
-import javax.servlet.ServletContextEvent;
-import javax.servlet.ServletContextListener;
-import javax.servlet.annotation.WebListener;
-
-@WebListener
-public class ContextLoaderListener implements ServletContextListener {
-  private static final Logger logger = LoggerFactory.getLogger(ContextLoaderListener.class);
-
-  @Override
-  public void contextInitialized(ServletContextEvent sce) {
-    ResourceDatabasePopulator populator = new ResourceDatabasePopulator();
-    populator.addScript(new ClassPathResource("schema.sql"));
-    populator.addScript(new ClassPathResource("data.sql"));
-    DatabasePopulatorUtils.execute(populator, ConnectionManager.getDatasource());
-
-    logger.info("Completed loading ServletContext");
-  }
-
-  @Override
-  public void contextDestroyed(ServletContextEvent sce) {
-  }
-}
